# swagger_server/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: Crypto Metrics Service
  description: This is a Simple Service to Store and Query Crypto Metrics.
  version: 1.0.0
servers:
- url: https://localhost:8080/
- url: https://localhost:8080/
...
paths:
...
  /monitors:
    get:
      tags:
      - Crypto Metrics App
      summary: list all monitors for a metric
      operationId: list_metric_monitors
      parameters:
      - name: metric_id
        in: query
        description: metric id to monitor
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Monitor'
                x-content-type: application/json
      security:
      - api_key: []
      x-openapi-router-controller: swagger_server.controllers.crypto_metrics_app_controller
    post:
      tags:
      - Crypto Metrics App
      summary: set a monitor for a metric
      operationId: set_metric_monitor
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Monitor'
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Monitor'
      security:
      - api_key: []
      x-openapi-router-controller: swagger_server.controllers.crypto_metrics_app_controller
  /monitors/{monitor_id}:
    get:
      tags:
      - Crypto Metrics App
      summary: deelte a monitor for the metric
      operationId: delete_metric_monitor
      parameters:
      - name: metric_id
        in: path
        description: metric id to monitor
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: monitor_id
        in: path
        description: metric id to monitor
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Monitor'
                x-content-type: application/json
      security:
      - api_key: []
      x-openapi-router-controller: swagger_server.controllers.crypto_metrics_app_controller
...
components:
  schemas:
    Monitor:
      type: object
      properties:
        monitor_id:
          type: string
        metric_id:
          type: string
        operand:
          type: string
          enum:
          - greater_than
          - less_than
        value:
          type: number
          format: float
        notify_webhook:
          type: string
      example:
        metric_id: metric_id
        monitor_id: monitor_id
        notify_webhook: notify_webhook
        value: 0.8008282
        operand: ""
...
  securitySchemes:
    api_key:
      type: apiKey
      name: api_key
      in: header
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_api_key